+++ CHANGELOG.md
@@ -34,6 +34,8 @@ All notable changes to this project will be documented in this file.
 ### Fixed ###
 - Signal 11 when connecting to [Slack](https://slack.com/).
 - Instead of using slots for chunks of names, use a linked list.
+- In the printtext module: switch off all terminal attributes during
+  indentation.
 
 ## [1.2] - 2016-09-22 ##
 ### Added ###